int > size_t

+++ src/Library.cpp
@@ -13,7 +13,7 @@ std::vector<float> Library::generateRandomValues(int amount,
   std::mt19937 gen(rd());
   std::uniform_real_distribution<float> dis(min, max);
   std::vector<float> randomValues(amount);
-  for (int i = 0; i < amount; ++i) {
+  for (size_t i = 0; i < amount; ++i) {
     randomValues[i] = dis(gen);
   }
   return randomValues;
